WordCamp Organizer Reminders: Refactor mail logic out of callers and into mail().
Some of the logic container in functions that call mail() is duplicated, and also is more appropriate to be handled
in mail() itself. This keeps the callers more lean and maintainable.

This will also make it easier to add extra types of callers, like one to manually send an email.