WordCamp Organizer Reminders: Get address from $_POST instead of database.
If the post is a draft and the email field is empty, then the admin enters the email and saves sets it to pending
in a single step, the previous code would have pulled an empty string from the database instead of the new address
from $_POST, and the e-mail would not have been sent.

Pulling from $_POST instead of the database ensures that the most recent value will always be retrieved.