WordCamp Participation Notifier: Split attendee_registered() into two functions.
Previously it was triggered when the buyer's username was saved, but that happens before the ticket is successfully
purchased, resulting in a bug where a notification would be sent even if the transaction eventually failed.

The notification shouldn't be sent until the buyer's attendee post has moved to the publish state. Consequentially,
that event is separate from the event when the additional attendees confirm their registration, so we now have
seprate methods to handle each event.