WordCamp Post Types: Prevent duplicate avatars on [speakers] page.
r933 and r935 added functionality to display extra information on single speaker/session posts, and attempted to make sure that it only appeared on those posts by checking `$post->post_type`.

That resulted in a bug where the extra info would also be added to regular pages that contained the [speakers] or [sessions] shortcodes, because in that context `$post` would refer to the current post in the secondary loop, rather than the page from the main loop.