WP.org Themes: Don't update theme post until new version was approved.
Theme description and tags might change in new versions, so the post can't be
updated with that new information until that version has been approved.

This also fixes a bug where the post date would be set to the upload date of
the new version, mixing new themes with updated themes in the `Latest Themes`
view in the Directory.

Fixes #943.