WordCamp.org: Centralize user fetching functionality.
Previously `wordcamp-participation-notifier` and `wc-post-types` were using slightly different code to get the same data, which resulted in a bug where `wc-post-types` would find the user, but `wordcamp-participation-notifier` wouldn't. That led to some profile badges being removed when the speaker/organizer post was updated, because the notifier would think the username was removed from the field, and send a command to remove the badge.

Because the two plugins are interdependent, they need to use a common function to guarantee identical results.

`get_saved_wporg_user_id()` doesn't need to use the new function, because `tix_username` is always set from `user_login`, rather than being entered as user-input.

See #977